requirements.txt
absl-py==2.1.0
anyio==4.5.2
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
attrs==25.3.0
babel==2.17.0
backcall @ file:///home/conda/feedstock_root/build_artifacts/backcall_1592338393461/work
beautifulsoup4==4.13.3
bleach==6.1.0
catkin-pkg==1.0.0
certifi==2024.7.4
cffi==1.17.1
chardet==4.0.0
charset-normalizer==3.3.2
comm @ file:///home/conda/feedstock_root/build_artifacts/comm_1710320294760/work
contourpy==1.1.1
cycler==0.12.1
d2l==0.17.6
debugpy @ file:///croot/debugpy_1690905042057/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1641555617451/work
defusedxml==0.7.1
-e git+https://github.com/ssssuxin/ALOHA-act.git@c658a35325b0a4ce6582812ab3473e70f1912a26#egg=detr&subdirectory=detr
distro==1.9.0
dm-control==1.0.14
dm-env==1.6
dm-tree==0.1.8
docutils==0.20.1
einops==0.8.0
entrypoints @ file:///home/conda/feedstock_root/build_artifacts/entrypoints_1643888246732/work
exceptiongroup==1.2.2
executing==2.0.1
fastjsonschema==2.21.1
filelock==3.15.4
fonttools==4.53.1
fqdn==1.5.1
fsspec==2024.6.1
glfw==2.7.0
h11==0.14.0
h5py==3.11.0
httpcore==1.0.7
httpx==0.28.1
idna==2.10
importlib_metadata==8.5.0
importlib_resources==6.4.4
ipykernel @ file:///home/conda/feedstock_root/build_artifacts/ipykernel_1719845459717/work
ipython @ file:///home/conda/feedstock_root/build_artifacts/ipython_1683289033986/work
ipython-genutils==0.2.0
ipywidgets==8.1.5
isoduration==20.11.0
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1696326070614/work
Jinja2==3.1.4
joblib==1.4.2
json5==0.10.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==7.4.9
jupyter_core @ file:///home/conda/feedstock_root/build_artifacts/jupyter_core_1727163409502/work
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.3.6
jupyterlab_pygments==0.3.0
jupyterlab_widgets==3.0.13
kiwisolver==1.4.5
labmaze==1.0.6
lxml==5.3.0
MarkupSafe==2.1.5
matplotlib==3.5.1
matplotlib-inline==0.1.6
mistune==3.1.3
mpmath==1.3.0
mujoco==2.3.7
nbclassic==1.2.0
nbclient==0.10.1
nbconvert==7.16.6
nbformat==5.10.4
nest_asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1705850609492/work
networkx==3.1
notebook==6.5.7
notebook_shim==0.2.4
numpy==1.21.5
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.20
nvidia-nvtx-cu12==12.1.105
onnx==1.16.2
opencv-python==4.10.0.84
overrides==7.7.0
packaging==24.1
pandas==1.2.4
pandocfilters==1.5.1
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1712320355065/work
pexpect @ file:///home/conda/feedstock_root/build_artifacts/pexpect_1706113125309/work
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
pillow==10.4.0
pkgutil_resolve_name==1.3.10
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/platformdirs_1726613481435/work
prometheus_client==0.21.1
prompt_toolkit==3.0.47
protobuf==5.27.3
psutil @ file:///home/conda/feedstock_root/build_artifacts/psutil_1653089172347/work
ptyprocess @ file:///home/conda/feedstock_root/build_artifacts/ptyprocess_1609419310487/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pure_eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1721585709575/work
py-cpuinfo==9.0.0
pycparser==2.22
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1714846767233/work
PyOpenGL==3.1.7
pyparsing==3.0.9
pyquaternion==0.9.9
pyrealsense2==2.55.1.6486
python-dateutil==2.9.0.post0
python-json-logger==3.3.0
pytz==2024.2
PyYAML==6.0.2
pyzmq @ file:///croot/pyzmq_1705605076900/work
qtconsole==5.6.1
QtPy==2.4.3
referencing==0.35.1
requests==2.25.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rospkg==1.5.1
rpds-py==0.20.1
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.13.2
Send2Trash==1.8.3
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
sniffio==1.3.1
soupsieve==2.6
stack-data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1669632077133/work
sympy==1.13.2
tensorrt==8.6.0
tensorrt-cu12==10.4.0
tensorrt-cu12-bindings==10.4.0
tensorrt-cu12-libs==10.4.0
terminado==0.18.1
threadpoolctl==3.5.0
timm==0.4.5
tinycss2==1.2.1
tomli==2.2.1
torch==2.4.0
torchinfo==1.8.0
torchsummary==1.5.1
torchvision==0.19.0
tornado==6.4.2
tqdm==4.66.5
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1713535121073/work
triton==3.0.0
types-python-dateutil==2.9.0.20241206
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1717802530399/work
tzdata==2024.2
ultralytics-thop==2.0.6
uri-template==1.3.0
urllib3==1.26.20
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1704731205417/work
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
widgetsnbextension==4.0.13
zipp==3.20.0
